use build_conv_layer to claim dcn in resnets
